+++ persistence-cassandra/src/it/scala/com/evolutiongaming/kafka/flow/CassandraSessionStub.scala
@@ -1,16 +1,16 @@
 package com.evolutiongaming.kafka.flow
 
+import cats.MonadThrow
 import cats.effect.concurrent.Ref
 import cats.syntax.all._
 import com.datastax.driver.core.Statement
-import com.evolutiongaming.catshelper.MonadThrowable
 import com.evolutiongaming.kafka.journal.eventual.cassandra.CassandraSession
 import com.evolutiongaming.sstream.Stream
 
 object CassandraSessionStub {
 
-  def alwaysFails[F[_]: MonadThrowable]: CassandraSession[F] = new CassandraSession[F] {
-    def fail[T]: F[T] = MonadThrowable[F].raiseError {
+  def alwaysFails[F[_]: MonadThrow]: CassandraSession[F] = new CassandraSession[F] {
+    def fail[T]: F[T] = MonadThrow[F].raiseError {
       new RuntimeException("CassandraSessionStub: always fails")
     }
     def prepare(query: String) = fail
@@ -18,12 +18,12 @@ object CassandraSessionStub {
     def unsafe = sys.error("CassandraSessionStub: no unsafe session")
   }
 
-  def injectFailures[F[_]: MonadThrowable](
+  def injectFailures[F[_]: MonadThrow](
     session: CassandraSession[F],
     failAfter: Ref[F, Int]
   ): CassandraSession[F] = new CassandraSession[F] {
 
-    def fail[T](query: String): F[T] = MonadThrowable[F].raiseError {
+    def fail[T](query: String): F[T] = MonadThrow[F].raiseError {
       new RuntimeException(s"CassandraSessionStub: failing after proper calls exhausted: $query")
     }