+++ persistence-cassandra/src/main/scala/com/evolutiongaming/kafka/flow/key/CassandraKeys.scala
@@ -1,11 +1,11 @@
 package com.evolutiongaming.kafka.flow.key
 
 import cats.Monad
+import cats.MonadThrow
 import cats.effect.Clock
 import cats.syntax.all._
 import com.evolutiongaming.cassandra.sync.CassandraSync
 import com.evolutiongaming.catshelper.ClockHelper._
-import com.evolutiongaming.catshelper.MonadThrowable
 import com.evolutiongaming.kafka.flow.KafkaKey
 import com.evolutiongaming.kafka.journal.eventual.cassandra.CassandraSession
 import com.evolutiongaming.kafka.journal.eventual.cassandra.SegmentNr
@@ -177,7 +177,7 @@ class CassandraKeys[F[_]: Monad: Fail: Clock](
 object CassandraKeys {
 
   /** Creates schema in Cassandra if not there yet */
-  def withSchema[F[_]: MonadThrowable: Clock](
+  def withSchema[F[_]: MonadThrow: Clock](
     session: CassandraSession[F],
     sync: CassandraSync[F]
   ): F[KeyDatabase[F, KafkaKey]] = {