+++ persistence-cassandra/src/main/scala/com/evolutiongaming/kafka/flow/snapshot/CassandraSnapshots.scala
@@ -1,12 +1,12 @@
 package com.evolutiongaming.kafka.flow.snapshot
 
 import cats.Monad
+import cats.MonadThrow
 import cats.effect.Clock
 import cats.syntax.all._
 import com.datastax.driver.core.Row
 import com.evolutiongaming.cassandra.sync.CassandraSync
 import com.evolutiongaming.catshelper.ClockHelper._
-import com.evolutiongaming.catshelper.MonadThrowable
 import com.evolutiongaming.kafka.flow.KafkaKey
 import com.evolutiongaming.kafka.flow.cassandra.CassandraCodecs._
 import com.evolutiongaming.kafka.journal.FromBytes
@@ -19,7 +19,7 @@ import com.evolutiongaming.scassandra.syntax._
 import com.evolutiongaming.skafka.Offset
 import scodec.bits.ByteVector
 
-class CassandraSnapshots[F[_]: MonadThrowable: Clock, T](
+class CassandraSnapshots[F[_]: MonadThrow: Clock, T](
   session: CassandraSession[F]
 )(implicit fromBytes: FromBytes[F, T], toBytes: ToBytes[F, T]) extends SnapshotDatabase[F, KafkaKey, KafkaSnapshot[T]] {
 
@@ -204,7 +204,7 @@ class CassandraSnapshots[F[_]: MonadThrowable: Clock, T](
 object CassandraSnapshots {
 
   /** Creates schema in Cassandra if not there yet */
-  def withSchema[F[_]: MonadThrowable: Clock, T](
+  def withSchema[F[_]: MonadThrow: Clock, T](
     session: CassandraSession[F],
     sync: CassandraSync[F]
   )(implicit fromBytes: FromBytes[F, T], toBytes: ToBytes[F, T]): F[SnapshotDatabase[F, KafkaKey, KafkaSnapshot[T]]] =